Create uistart.sh
Bash script with accompanying icon that will start the node.js server for the UI on the base station. Checks to see if its running and if not starts it and opens a browser to the url. If an instance of the server is already running, a browser window will be opened to the url.